refactor: 역할에 맞도록 패키지 이동
