+++ app/src/main/java/woowacourse/omok/adapter/RuleResult.kt
@@ -1,6 +1,6 @@
 package woowacourse.omok.adapter
 
-import woowacourse.omok.domain.model.position.Stone
+import woowacourse.omok.domain.model.stone.Stone
 
 sealed class RuleResult {
     data class OnRule(val stone: Stone) : RuleResult()
