+++ app/src/main/java/woowacourse/omok/data/StoneRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package woowacourse.omok.data
 
-import woowacourse.omok.domain.model.StoneRepository
-import woowacourse.omok.domain.model.position.Stone
+import woowacourse.omok.domain.model.stone.Stone
 import woowacourse.omok.domain.model.stone.StoneType
 import woowacourse.omok.domain.model.stone.Stones
+import woowacourse.omok.domain.repository.StoneRepository
 
 class StoneRepositoryImpl(private val stoneDao: StoneDao) : StoneRepository {
     override fun insert(stone: Stone) {
