+++ app/src/main/java/woowacourse/omok/domain/model/rule/RenjuRule.kt
@@ -1,7 +1,7 @@
 package woowacourse.omok.domain.model.rule
 
 import woowacourse.omok.adapter.RuleResult
-import woowacourse.omok.domain.model.position.Stone
+import woowacourse.omok.domain.model.stone.Stone
 import woowacourse.omok.domain.model.stone.Stones
 
 interface RenjuRule {
