+++ app/src/main/java/woowacourse/ui/PlayEvent.kt
@@ -2,7 +2,7 @@ package woowacourse.ui
 
 import woowacourse.omok.adapter.RuleResult
 import woowacourse.omok.domain.model.position.Position
-import woowacourse.omok.domain.model.position.Stone
+import woowacourse.omok.domain.model.stone.Stone
 import woowacourse.omok.domain.model.stone.StoneType
 
 interface PlayEvent {
